requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.2
minio>=7.1.13
pydantic>=2.0.0
python-multipart>=0.0.6
GPUtil>=1.4.0
torch>=2.0.0
redis>=4.5.5
Pillow>=10.0.0
PyMuPDF>=1.23.0
transformers>=4.30.0
qwen-vl-utils>=0.0.1
requests>=2.31.0
